feat: bundle prebuilt PJSIP (pjsua2) and port to real bindings (#5)
* feat: bundle prebuilt PJSIP (pjsua2) and port to real bindings

Build PJSIP 2.16 + pjsua2 (G.711/G.722/Opus, OpenSSL/TLS) via
pjsip-android-builder and commit the artifacts so the app builds with SIP
enabled out of the box: pjsua2.jar plus libpjsua2.so + libc++_shared.so for
arm64-v8a, armeabi-v7a and x86_64 (native .so stored in Git LFS).

Port the app and the compile-time stub from the hand-written enum API
(MediaFrameType, .swigValue()/.swigToEnum()) to real pjsua2's int-constant
enums, so both the real and stub build paths compile. Unit tests pass.


* fix: run unit tests against the stub with real PJSIP committed; document setup

The committed pjsua2.jar made CI use the real SWIG bindings for unit tests,
which try to System.loadLibrary("pjsua2") on the JVM and fail with
UnsatisfiedLinkError (the native .so can't load in JVM/Robolectric tests).

- Compile & run unit tests against the stub even when the real jar is present,
  and filter the jar off the unit-test classpath (build.gradle.kts).
- Make the unit-tested stub classes (ByteVector, MediaFrame, MediaFormat,
  MediaFormatAudio) binary-compatible with real SWIG (getter/setter shape,
  explicit ByteVector.add(Short)/get(int)) so app code compiled against the jar
  resolves against the stub at test runtime.
- CI: pull Git LFS (lfs: true) in the APK-assembling jobs so the debug/release
  APKs ship real native libs instead of pointer stubs.
- Docs: README + docs/PJSIP_BUILD.md now cover cloning with Git LFS, what's
  bundled, building/testing, and how to rebuild PJSIP.


---------

+++ app/src/pjsipStub/java/org/pjsip/pjsua2/Call.java
@@ -18,7 +18,7 @@ public class Call {
     public void answer(CallOpParam prm) throws Exception {}
     public void hangup(CallOpParam prm) throws Exception {}
     public void makeCall(String dst, CallOpParam prm) throws Exception {}
-    public AudioMedia getAudioMedia(long medIdx) throws Exception { return new AudioMedia(); }
+    public AudioMedia getAudioMedia(int medIdx) throws Exception { return new AudioMedia(); }
     public void delete() {}
     public void onCallState(OnCallStateParam prm) {}
     public void onCallMediaState(OnCallMediaStateParam prm) {}
